Refactor UnitItem component file structure
